Obtain a new KeyContext if the old one expires or runs out of encryptions
When obtaining a future through getKeyContext, if the cached future is done we
check if it can handle the next batch of encryptions, if not we request a new
keyContext and cache the new future.

Why:
We need to change to a new DEK because by the NIST guidelines we should only
use the key material to encrypt some limited amount of data and it should also
be bound by time.