Switch back to filter dispatch thread after KMS operations
Note: the FilterThreadExecutor could be a candidate for extracting
an interface and pulling up to the FilterFactoryContext as a general
facility (it would likely extend ScheduledExecutorService and also
off an `boolean isInFilterThread()` method for Filter Authors.

Why:
We are working with non-threadsafe structures and so want to be sure
that access is limited to the eventloop of the filter instance.

Signed-off-by: Robert Young <robeyoun@redhat.com>